TEST: try to reintroduce an LRU cache between rocksdb and runtime
This is a very crude quick fix/hack. I tried to cache
`UniversalExecutable`s, but that led to a swarm of bats coming outta my
nose. We cannot cache `UniversalArtifacts` or anything of the sort as
they are not shareable across threads (and are locked to the specific
runtime instance they're loaded with).

https://near.zulipchat.com/#narrow/stream/313099-pagoda.2Fstorage/topic/background.20node.20fetching/near/385948525
suggests that this should improve the situation slightly at least in
mocknet test?

I don't think this should go into production though, but it could
inform us of some changes we could try making to our storage? For
example – we could look at accessing the contracts that we expect to be
part of the upcoming chunk before we start processing the chunk, which
would help masking the flat performance hit we're seeing.